This class is a generic container for the model used for outbreak detection

This class is a generic container for the model used for outbreak detection

Slots

nStates

Number of states in the hidden Markov model.

emission

Emission of hidden Markov model. Includes a excodeFamily and a excodeFormula object.

transitions

Transition proabilities of hidden Markov model.

transitions_prior

Indicates whether a prior for transition probabilities should be used.

prior_weights

Dirichlet prior weights for transition probabilities. This cannot be changed by the user.

loglik_transitions

Prior log likelihood of transition probabilities. Only relevant for model fitting.

initial_prob

Initial state probabilities of the hidden Markov model.

setBckgState

Indicates whether a background state should be computed for model fitting. Background states are set to 0s and time points where the Anscombe residuals of the initial model are < 1.

id

Name of the time series

method

Method for parameter estimation, one of c("supervised", "unsupervised"). Default: "unsupervised".

converged

Logical, indicating whether the EM-algorithm converged.

niter

Maximum number of iterations for model fitting.

LogLik

Log lokelihood of the fitted model.

AIC

Akaike information criterion of the fitted model.

BIC

Bayesian information criterion of the fitted model.

posterior

Matrix of posterior probabilities for each time point (rows) and state (columns).

alpha

Alpha (Forward) probabilities of the HMM.

pval

P-value for testing wheter timepoint is in normal or excess state.

timepoint_fit

Numeric time point in timeseries used for fitting.

timepoint

Numeric time point in timeseries used for fitting.

date

Date for each time point.

observed

Vector of observed number of cases.

population

Population size, default: 1.

error

Character string of error message (if an error occured).